Add VM Evaluation Functions for Array and Struct Literals
- Introduced new functions in the VM for evaluating array and struct literals, enhancing the ability to handle complex data types.
- Implemented error handling for nil literals and type mismatches, ensuring robust evaluation processes.
- Added support for field access in structs and index operations in arrays, improving the VM's capability to manipulate structured data.
- Enhanced memory management by allocating arrays and structs in the heap, facilitating efficient data handling during execution.
- Updated existing evaluation logic to accommodate new data types, ensuring seamless integration with the VM's execution flow.
