+++ internal/vm/eval.go
@@ -2,12 +2,10 @@ package vm
 
 import (
 	"fmt"
-	"math"
 	"strings"
 
 	"fortio.org/safecast"
 
-	"surge/internal/ast"
 	"surge/internal/mir"
 	"surge/internal/types"
 	"surge/internal/vm/bignum"
@@ -212,740 +210,6 @@ func (vm *VM) evalConst(c *mir.Const) Value {
 		return MakeNothing()
 	default:
 		return Value{Kind: VKInvalid}
-	}
-}
-
-// evalBinaryOp evaluates a binary operation.
-func (vm *VM) evalBinaryOp(op ast.ExprBinaryOp, left, right Value) (Value, *VMError) {
-	switch op {
-	case ast.ExprBinaryAdd:
-		switch {
-		case left.Kind == VKBigInt && right.Kind == VKBigInt:
-			a, vmErr := vm.mustBigInt(left)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			b, vmErr := vm.mustBigInt(right)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			res, err := bignum.IntAdd(a, b)
-			if err != nil {
-				return Value{}, vm.bignumErr(err)
-			}
-			return vm.makeBigInt(left.TypeID, res), nil
-		case left.Kind == VKBigUint && right.Kind == VKBigUint:
-			a, vmErr := vm.mustBigUint(left)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			b, vmErr := vm.mustBigUint(right)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			res, err := bignum.UintAdd(a, b)
-			if err != nil {
-				return Value{}, vm.bignumErr(err)
-			}
-			return vm.makeBigUint(left.TypeID, res), nil
-		case left.Kind == VKBigFloat && right.Kind == VKBigFloat:
-			a, vmErr := vm.mustBigFloat(left)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			b, vmErr := vm.mustBigFloat(right)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			res, err := bignum.FloatAdd(a, b)
-			if err != nil {
-				return Value{}, vm.bignumErr(err)
-			}
-			return vm.makeBigFloat(left.TypeID, res), nil
-		case left.Kind == VKInt && right.Kind == VKInt:
-			res, ok := AddInt64Checked(left.Int, right.Int)
-			if !ok {
-				return Value{}, vm.eb.intOverflow()
-			}
-			return MakeInt(res, left.TypeID), nil
-		default:
-			return Value{}, vm.eb.typeMismatch("numeric", fmt.Sprintf("%s and %s", left.Kind, right.Kind))
-		}
-
-	case ast.ExprBinarySub:
-		switch {
-		case left.Kind == VKBigInt && right.Kind == VKBigInt:
-			a, vmErr := vm.mustBigInt(left)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			b, vmErr := vm.mustBigInt(right)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			res, err := bignum.IntSub(a, b)
-			if err != nil {
-				return Value{}, vm.bignumErr(err)
-			}
-			return vm.makeBigInt(left.TypeID, res), nil
-		case left.Kind == VKBigUint && right.Kind == VKBigUint:
-			a, vmErr := vm.mustBigUint(left)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			b, vmErr := vm.mustBigUint(right)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			res, err := bignum.UintSub(a, b)
-			if err != nil {
-				return Value{}, vm.bignumErr(err)
-			}
-			return vm.makeBigUint(left.TypeID, res), nil
-		case left.Kind == VKBigFloat && right.Kind == VKBigFloat:
-			a, vmErr := vm.mustBigFloat(left)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			b, vmErr := vm.mustBigFloat(right)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			res, err := bignum.FloatSub(a, b)
-			if err != nil {
-				return Value{}, vm.bignumErr(err)
-			}
-			return vm.makeBigFloat(left.TypeID, res), nil
-		case left.Kind == VKInt && right.Kind == VKInt:
-			res, ok := SubInt64Checked(left.Int, right.Int)
-			if !ok {
-				return Value{}, vm.eb.intOverflow()
-			}
-			return MakeInt(res, left.TypeID), nil
-		default:
-			return Value{}, vm.eb.typeMismatch("numeric", fmt.Sprintf("%s and %s", left.Kind, right.Kind))
-		}
-
-	case ast.ExprBinaryMul:
-		switch {
-		case left.Kind == VKBigInt && right.Kind == VKBigInt:
-			a, vmErr := vm.mustBigInt(left)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			b, vmErr := vm.mustBigInt(right)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			res, err := bignum.IntMul(a, b)
-			if err != nil {
-				return Value{}, vm.bignumErr(err)
-			}
-			return vm.makeBigInt(left.TypeID, res), nil
-		case left.Kind == VKBigUint && right.Kind == VKBigUint:
-			a, vmErr := vm.mustBigUint(left)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			b, vmErr := vm.mustBigUint(right)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			res, err := bignum.UintMul(a, b)
-			if err != nil {
-				return Value{}, vm.bignumErr(err)
-			}
-			return vm.makeBigUint(left.TypeID, res), nil
-		case left.Kind == VKBigFloat && right.Kind == VKBigFloat:
-			a, vmErr := vm.mustBigFloat(left)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			b, vmErr := vm.mustBigFloat(right)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			res, err := bignum.FloatMul(a, b)
-			if err != nil {
-				return Value{}, vm.bignumErr(err)
-			}
-			return vm.makeBigFloat(left.TypeID, res), nil
-		case left.Kind == VKInt && right.Kind == VKInt:
-			res, ok := MulInt64Checked(left.Int, right.Int)
-			if !ok {
-				return Value{}, vm.eb.intOverflow()
-			}
-			return MakeInt(res, left.TypeID), nil
-		default:
-			return Value{}, vm.eb.typeMismatch("numeric", fmt.Sprintf("%s and %s", left.Kind, right.Kind))
-		}
-
-	case ast.ExprBinaryDiv:
-		switch {
-		case left.Kind == VKBigInt && right.Kind == VKBigInt:
-			a, vmErr := vm.mustBigInt(left)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			b, vmErr := vm.mustBigInt(right)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			q, _, err := bignum.IntDivMod(a, b)
-			if err != nil {
-				return Value{}, vm.bignumErr(err)
-			}
-			return vm.makeBigInt(left.TypeID, q), nil
-		case left.Kind == VKBigUint && right.Kind == VKBigUint:
-			a, vmErr := vm.mustBigUint(left)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			b, vmErr := vm.mustBigUint(right)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			q, _, err := bignum.UintDivMod(a, b)
-			if err != nil {
-				return Value{}, vm.bignumErr(err)
-			}
-			return vm.makeBigUint(left.TypeID, q), nil
-		case left.Kind == VKBigFloat && right.Kind == VKBigFloat:
-			a, vmErr := vm.mustBigFloat(left)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			b, vmErr := vm.mustBigFloat(right)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			res, err := bignum.FloatDiv(a, b)
-			if err != nil {
-				return Value{}, vm.bignumErr(err)
-			}
-			return vm.makeBigFloat(left.TypeID, res), nil
-		case left.Kind == VKInt && right.Kind == VKInt:
-			if right.Int == 0 {
-				return Value{}, vm.eb.divisionByZero()
-			}
-			return MakeInt(left.Int/right.Int, left.TypeID), nil
-		default:
-			return Value{}, vm.eb.typeMismatch("numeric", fmt.Sprintf("%s and %s", left.Kind, right.Kind))
-		}
-
-	case ast.ExprBinaryMod:
-		switch {
-		case left.Kind == VKBigInt && right.Kind == VKBigInt:
-			a, vmErr := vm.mustBigInt(left)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			b, vmErr := vm.mustBigInt(right)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			_, r, err := bignum.IntDivMod(a, b)
-			if err != nil {
-				return Value{}, vm.bignumErr(err)
-			}
-			return vm.makeBigInt(left.TypeID, r), nil
-		case left.Kind == VKBigUint && right.Kind == VKBigUint:
-			a, vmErr := vm.mustBigUint(left)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			b, vmErr := vm.mustBigUint(right)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			_, r, err := bignum.UintDivMod(a, b)
-			if err != nil {
-				return Value{}, vm.bignumErr(err)
-			}
-			return vm.makeBigUint(left.TypeID, r), nil
-		case left.Kind == VKInt && right.Kind == VKInt:
-			if right.Int == 0 {
-				return Value{}, vm.eb.divisionByZero()
-			}
-			return MakeInt(left.Int%right.Int, left.TypeID), nil
-		default:
-			return Value{}, vm.eb.typeMismatch("numeric", fmt.Sprintf("%s and %s", left.Kind, right.Kind))
-		}
-
-	case ast.ExprBinaryEq:
-		if left.Kind != right.Kind {
-			return Value{}, vm.eb.typeMismatch(left.Kind.String(), right.Kind.String())
-		}
-		var result bool
-		switch left.Kind {
-		case VKInt:
-			result = left.Int == right.Int
-		case VKBigInt:
-			a, vmErr := vm.mustBigInt(left)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			b, vmErr := vm.mustBigInt(right)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			result = a.Cmp(b) == 0
-		case VKBigUint:
-			a, vmErr := vm.mustBigUint(left)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			b, vmErr := vm.mustBigUint(right)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			result = a.Cmp(b) == 0
-		case VKBigFloat:
-			a, vmErr := vm.mustBigFloat(left)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			b, vmErr := vm.mustBigFloat(right)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			result = a.Cmp(b) == 0
-		case VKBool:
-			result = left.Bool == right.Bool
-		case VKHandleString:
-			lObj := vm.Heap.Get(left.H)
-			rObj := vm.Heap.Get(right.H)
-			if lObj == nil || rObj == nil {
-				return Value{}, vm.eb.makeError(PanicOutOfBounds, "invalid string handle")
-			}
-			result = lObj.Str == rObj.Str
-		default:
-			result = left.H == right.H
-		}
-		return MakeBool(result, types.NoTypeID), nil
-
-	case ast.ExprBinaryNotEq:
-		if left.Kind != right.Kind {
-			return Value{}, vm.eb.typeMismatch(left.Kind.String(), right.Kind.String())
-		}
-		var result bool
-		switch left.Kind {
-		case VKInt:
-			result = left.Int != right.Int
-		case VKBigInt:
-			a, vmErr := vm.mustBigInt(left)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			b, vmErr := vm.mustBigInt(right)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			result = a.Cmp(b) != 0
-		case VKBigUint:
-			a, vmErr := vm.mustBigUint(left)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			b, vmErr := vm.mustBigUint(right)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			result = a.Cmp(b) != 0
-		case VKBigFloat:
-			a, vmErr := vm.mustBigFloat(left)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			b, vmErr := vm.mustBigFloat(right)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			result = a.Cmp(b) != 0
-		case VKBool:
-			result = left.Bool != right.Bool
-		case VKHandleString:
-			lObj := vm.Heap.Get(left.H)
-			rObj := vm.Heap.Get(right.H)
-			if lObj == nil || rObj == nil {
-				return Value{}, vm.eb.makeError(PanicOutOfBounds, "invalid string handle")
-			}
-			result = lObj.Str != rObj.Str
-		default:
-			result = left.H != right.H
-		}
-		return MakeBool(result, types.NoTypeID), nil
-
-	case ast.ExprBinaryLess:
-		switch {
-		case left.Kind == VKBigInt && right.Kind == VKBigInt:
-			a, vmErr := vm.mustBigInt(left)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			b, vmErr := vm.mustBigInt(right)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			return MakeBool(a.Cmp(b) < 0, types.NoTypeID), nil
-		case left.Kind == VKBigUint && right.Kind == VKBigUint:
-			a, vmErr := vm.mustBigUint(left)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			b, vmErr := vm.mustBigUint(right)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			return MakeBool(a.Cmp(b) < 0, types.NoTypeID), nil
-		case left.Kind == VKBigFloat && right.Kind == VKBigFloat:
-			a, vmErr := vm.mustBigFloat(left)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			b, vmErr := vm.mustBigFloat(right)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			return MakeBool(a.Cmp(b) < 0, types.NoTypeID), nil
-		case left.Kind == VKInt && right.Kind == VKInt:
-			return MakeBool(left.Int < right.Int, types.NoTypeID), nil
-		default:
-			return Value{}, vm.eb.typeMismatch("numeric", fmt.Sprintf("%s and %s", left.Kind, right.Kind))
-		}
-
-	case ast.ExprBinaryLessEq:
-		switch {
-		case left.Kind == VKBigInt && right.Kind == VKBigInt:
-			a, vmErr := vm.mustBigInt(left)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			b, vmErr := vm.mustBigInt(right)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			return MakeBool(a.Cmp(b) <= 0, types.NoTypeID), nil
-		case left.Kind == VKBigUint && right.Kind == VKBigUint:
-			a, vmErr := vm.mustBigUint(left)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			b, vmErr := vm.mustBigUint(right)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			return MakeBool(a.Cmp(b) <= 0, types.NoTypeID), nil
-		case left.Kind == VKBigFloat && right.Kind == VKBigFloat:
-			a, vmErr := vm.mustBigFloat(left)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			b, vmErr := vm.mustBigFloat(right)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			return MakeBool(a.Cmp(b) <= 0, types.NoTypeID), nil
-		case left.Kind == VKInt && right.Kind == VKInt:
-			return MakeBool(left.Int <= right.Int, types.NoTypeID), nil
-		default:
-			return Value{}, vm.eb.typeMismatch("numeric", fmt.Sprintf("%s and %s", left.Kind, right.Kind))
-		}
-
-	case ast.ExprBinaryGreater:
-		switch {
-		case left.Kind == VKBigInt && right.Kind == VKBigInt:
-			a, vmErr := vm.mustBigInt(left)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			b, vmErr := vm.mustBigInt(right)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			return MakeBool(a.Cmp(b) > 0, types.NoTypeID), nil
-		case left.Kind == VKBigUint && right.Kind == VKBigUint:
-			a, vmErr := vm.mustBigUint(left)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			b, vmErr := vm.mustBigUint(right)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			return MakeBool(a.Cmp(b) > 0, types.NoTypeID), nil
-		case left.Kind == VKBigFloat && right.Kind == VKBigFloat:
-			a, vmErr := vm.mustBigFloat(left)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			b, vmErr := vm.mustBigFloat(right)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			return MakeBool(a.Cmp(b) > 0, types.NoTypeID), nil
-		case left.Kind == VKInt && right.Kind == VKInt:
-			return MakeBool(left.Int > right.Int, types.NoTypeID), nil
-		default:
-			return Value{}, vm.eb.typeMismatch("numeric", fmt.Sprintf("%s and %s", left.Kind, right.Kind))
-		}
-
-	case ast.ExprBinaryGreaterEq:
-		switch {
-		case left.Kind == VKBigInt && right.Kind == VKBigInt:
-			a, vmErr := vm.mustBigInt(left)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			b, vmErr := vm.mustBigInt(right)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			return MakeBool(a.Cmp(b) >= 0, types.NoTypeID), nil
-		case left.Kind == VKBigUint && right.Kind == VKBigUint:
-			a, vmErr := vm.mustBigUint(left)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			b, vmErr := vm.mustBigUint(right)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			return MakeBool(a.Cmp(b) >= 0, types.NoTypeID), nil
-		case left.Kind == VKBigFloat && right.Kind == VKBigFloat:
-			a, vmErr := vm.mustBigFloat(left)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			b, vmErr := vm.mustBigFloat(right)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			return MakeBool(a.Cmp(b) >= 0, types.NoTypeID), nil
-		case left.Kind == VKInt && right.Kind == VKInt:
-			return MakeBool(left.Int >= right.Int, types.NoTypeID), nil
-		default:
-			return Value{}, vm.eb.typeMismatch("numeric", fmt.Sprintf("%s and %s", left.Kind, right.Kind))
-		}
-
-	default:
-		return Value{}, vm.eb.unimplemented(fmt.Sprintf("binary op %s", op))
-	}
-}
-
-// evalUnaryOp evaluates a unary operation.
-func (vm *VM) evalUnaryOp(op ast.ExprUnaryOp, operand Value) (Value, *VMError) {
-	switch op {
-	case ast.ExprUnaryMinus:
-		switch operand.Kind {
-		case VKBigInt:
-			i, vmErr := vm.mustBigInt(operand)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			return vm.makeBigInt(operand.TypeID, i.Negated()), nil
-		case VKBigFloat:
-			f, vmErr := vm.mustBigFloat(operand)
-			if vmErr != nil {
-				return Value{}, vmErr
-			}
-			return vm.makeBigFloat(operand.TypeID, bignum.FloatNeg(f)), nil
-		case VKInt:
-			if operand.Int == math.MinInt64 {
-				return Value{}, vm.eb.intOverflow()
-			}
-			return MakeInt(-operand.Int, operand.TypeID), nil
-		default:
-			return Value{}, vm.eb.typeMismatch("numeric", operand.Kind.String())
-		}
-
-	case ast.ExprUnaryNot:
-		if operand.Kind != VKBool {
-			return Value{}, vm.eb.typeMismatch("bool", operand.Kind.String())
-		}
-		return MakeBool(!operand.Bool, operand.TypeID), nil
-
-	case ast.ExprUnaryPlus:
-		switch operand.Kind {
-		case VKBigInt, VKBigUint, VKBigFloat, VKInt:
-			return operand, nil
-		default:
-			return Value{}, vm.eb.typeMismatch("numeric", operand.Kind.String())
-		}
-
-	case ast.ExprUnaryDeref:
-		switch operand.Kind {
-		case VKRef, VKRefMut:
-			return vm.loadLocationRaw(operand.Loc)
-		default:
-			return Value{}, vm.eb.derefOnNonRef(operand.Kind.String())
-		}
-
-	default:
-		return Value{}, vm.eb.unimplemented(fmt.Sprintf("unary op %s", op))
-	}
-}
-
-func (vm *VM) evalArrayLit(frame *Frame, lit *mir.ArrayLit) (Value, *VMError) {
-	if lit == nil {
-		return Value{}, vm.eb.makeError(PanicUnimplemented, "nil array literal")
-	}
-	elems := make([]Value, 0, len(lit.Elems))
-	for i := range lit.Elems {
-		v, vmErr := vm.evalOperand(frame, &lit.Elems[i])
-		if vmErr != nil {
-			return Value{}, vmErr
-		}
-		elems = append(elems, v)
-	}
-
-	h := vm.Heap.AllocArray(types.NoTypeID, elems)
-	return MakeHandleArray(h, types.NoTypeID), nil
-}
-
-// evalIndex evaluates an index operation.
-func (vm *VM) evalIndex(obj, idx Value) (Value, *VMError) {
-	maxIndex := int(^uint(0) >> 1)
-	maxInt := int64(maxIndex)
-	maxUint := uint64(^uint(0) >> 1)
-	var index int
-	switch idx.Kind {
-	case VKInt:
-		if idx.Int < 0 || idx.Int > maxInt {
-			return Value{}, vm.eb.outOfBounds(maxIndex, 0)
-		}
-		n, err := safecast.Conv[int](idx.Int)
-		if err != nil {
-			return Value{}, vm.eb.outOfBounds(maxIndex, 0)
-		}
-		index = n
-	case VKBigInt:
-		i, vmErr := vm.mustBigInt(idx)
-		if vmErr != nil {
-			return Value{}, vmErr
-		}
-		n, ok := i.Int64()
-		if !ok || n < 0 || n > maxInt {
-			return Value{}, vm.eb.outOfBounds(maxIndex, 0)
-		}
-		ni, err := safecast.Conv[int](n)
-		if err != nil {
-			return Value{}, vm.eb.outOfBounds(maxIndex, 0)
-		}
-		index = ni
-	case VKBigUint:
-		u, vmErr := vm.mustBigUint(idx)
-		if vmErr != nil {
-			return Value{}, vmErr
-		}
-		n, ok := u.Uint64()
-		if !ok || n > maxUint {
-			return Value{}, vm.eb.outOfBounds(maxIndex, 0)
-		}
-		ni, err := safecast.Conv[int](n)
-		if err != nil {
-			return Value{}, vm.eb.outOfBounds(maxIndex, 0)
-		}
-		index = ni
-	default:
-		return Value{}, vm.eb.typeMismatch("int", idx.Kind.String())
-	}
-
-	if obj.Kind != VKHandleArray {
-		return Value{}, vm.eb.typeMismatch("array", obj.Kind.String())
-	}
-	arrObj := vm.Heap.Get(obj.H)
-	if arrObj == nil {
-		return Value{}, vm.eb.makeError(PanicOutOfBounds, "invalid array handle")
-	}
-	if arrObj.Kind != OKArray {
-		return Value{}, vm.eb.makeError(PanicTypeMismatch, fmt.Sprintf("expected array handle, got %v", arrObj.Kind))
-	}
-	if index < 0 || index >= len(arrObj.Arr) {
-		return Value{}, vm.eb.outOfBounds(index, len(arrObj.Arr))
-	}
-	return vm.cloneForShare(arrObj.Arr[index])
-}
-
-func (vm *VM) evalStructLit(frame *Frame, lit *mir.StructLit) (Value, *VMError) {
-	if lit == nil {
-		return Value{}, vm.eb.makeError(PanicUnimplemented, "nil struct literal")
-	}
-	layout, vmErr := vm.layouts.Struct(lit.TypeID)
-	if vmErr != nil {
-		return Value{}, vmErr
-	}
-	fields := make([]Value, len(layout.FieldNames))
-	for i := range fields {
-		fields[i] = Value{Kind: VKInvalid}
-	}
-	for i := range lit.Fields {
-		f := &lit.Fields[i]
-		val, vmErr := vm.evalOperand(frame, &f.Value)
-		if vmErr != nil {
-			return Value{}, vmErr
-		}
-		idx, ok := layout.IndexByName[f.Name]
-		if !ok {
-			return Value{}, vm.eb.makeError(PanicTypeMismatch, fmt.Sprintf("struct type#%d has no field %q", layout.TypeID, f.Name))
-		}
-		fields[idx] = val
-	}
-	h := vm.Heap.AllocStruct(layout.TypeID, fields)
-	return MakeHandleStruct(h, lit.TypeID), nil
-}
-
-func (vm *VM) evalFieldAccess(frame *Frame, fa *mir.FieldAccess) (Value, *VMError) {
-	if fa == nil {
-		return Value{}, vm.eb.makeError(PanicUnimplemented, "nil field access")
-	}
-	if fa.FieldIdx >= 0 {
-		return Value{}, vm.eb.unimplemented("tuple field access")
-	}
-	if fa.FieldName == "" {
-		return Value{}, vm.eb.makeError(PanicTypeMismatch, "missing field name")
-	}
-	obj, vmErr := vm.evalOperand(frame, &fa.Object)
-	if vmErr != nil {
-		return Value{}, vmErr
-	}
-	if obj.Kind != VKHandleStruct {
-		return Value{}, vm.eb.typeMismatch("struct", obj.Kind.String())
-	}
-	sobj := vm.Heap.Get(obj.H)
-	if sobj == nil {
-		return Value{}, vm.eb.makeError(PanicOutOfBounds, "invalid struct handle")
-	}
-	if sobj.Kind != OKStruct {
-		return Value{}, vm.eb.typeMismatch("struct", fmt.Sprintf("%v", sobj.Kind))
-	}
-	layout, vmErr := vm.layouts.Struct(sobj.TypeID)
-	if vmErr != nil {
-		return Value{}, vmErr
-	}
-	idx, ok := layout.IndexByName[fa.FieldName]
-	if !ok {
-		return Value{}, vm.eb.makeError(PanicOutOfBounds, fmt.Sprintf("unknown field %q on type#%d", fa.FieldName, sobj.TypeID))
-	}
-	if idx < 0 || idx >= len(sobj.Fields) {
-		return Value{}, vm.eb.makeError(PanicOutOfBounds, fmt.Sprintf("field index %d out of bounds for type#%d", idx, sobj.TypeID))
-	}
-	return vm.cloneForShare(sobj.Fields[idx])
-}
-
-func (vm *VM) cloneForShare(v Value) (Value, *VMError) {
-	switch v.Kind {
-	case VKHandleString:
-		obj := vm.Heap.Get(v.H)
-		if obj == nil {
-			return Value{}, vm.eb.makeError(PanicOutOfBounds, "invalid string handle")
-		}
-		h := vm.Heap.AllocString(obj.TypeID, obj.Str)
-		return MakeHandleString(h, obj.TypeID), nil
-	case VKHandleArray, VKHandleStruct:
-		return Value{}, vm.eb.unimplemented(fmt.Sprintf("cloning %s", v.Kind))
-	default:
-		return v, nil
 	}
 }
